Rename and move prepareHome
